chore: remove local spec artifacts
