--- specs/sc-112048-fix-date-trunc-grouping/checklists/requirements.md
@@ -1,35 +1,0 @@
-# Specification Quality Checklist: Preserve Equivalent Time-Grain Expressions
-
-**Purpose**: Validate specification completeness and quality before proceeding to planning
-**Created**: 2026-07-15
-**Feature**: [spec.md](../spec.md)
-
-## Content Quality
-
-- [x] No implementation details (languages, frameworks, APIs)
-- [x] Focused on user value and business needs
-- [x] Written for non-technical stakeholders
-- [x] All mandatory sections completed
-
-## Requirement Completeness
-
-- [x] No [NEEDS CLARIFICATION] markers remain
-- [x] Requirements are testable and unambiguous
-- [x] Success criteria are measurable
-- [x] Success criteria are technology-agnostic (no implementation details)
-- [x] All acceptance scenarios are defined
-- [x] Edge cases are identified
-- [x] Scope is clearly bounded
-- [x] Dependencies and assumptions identified
-
-## Feature Readiness
-
-- [x] All functional requirements have clear acceptance criteria
-- [x] User scenarios cover primary flows
-- [x] Feature meets measurable outcomes defined in Success Criteria
-- [x] No implementation details leak into specification
-
-## Notes
-
-- Validation completed in one pass. No clarification markers remain.
-- Database product names and SQL concepts are retained as observable compatibility requirements from the source ticket, not as implementation prescriptions.
